requirements.txt
streamlit>=1.32.0
pandas>=2.2.3
plotly>=5.24.0
Pillow>=10.4.0

